+++ Changelog.md
@@ -3,6 +3,8 @@
 ### Fixed
     * Fix errors after trying to edit ame of product to exists name
     * Fix errors in console after click update after open/close result_set list
+    * Fix errors after change run name with existed run name
+    * Fix errors after change plan name with existed plan name
 ## [0.3.3]
 ### Fixed
     * Blocked status visible in add_result statuses